Port 2015 D1

+++ aoc-15/src/lib.rs
@@ -2,7 +2,7 @@ use std::str;
 
 mod day_01;
 
-pub fn solve(day: aoc::Day, part: aoc::Part, input: &str) -> Result<usize, aoc::Error> {
+pub fn solve(day: aoc::Day, part: aoc::Part, input: &str) -> Result<i32, aoc::Error> {
     use aoc::Parse;
     match day {
     | aoc::Day::D01 => day_01::NotQuiteLisp::parse(input),
